arreglando colors
