+++ front/src/pages/dashboard/index.js
@@ -1,10 +1,8 @@
 export * from "@/pages/dashboard/home";
-export * from "@/pages/dashboard/profile";
-export * from "@/pages/dashboard/tables";
 export * from "@/pages/dashboard/notifications";
 export * from "@/pages/dashboard/modules";
 export * from "@/pages/dashboard/team";
-export * from "@/pages/dashboard/profileNew";
+export * from "@/pages/dashboard/profile";
 export * from "@/pages/dashboard/wellnessHub";
 export * from "@/pages/dashboard/ownTracking";
 export * from "@/pages/dashboard/teams";
